Upload Packs

+++ src/Pack.java
@@ -8,13 +8,13 @@ public class Pack {
 	//Depending on the type. Rolled = weight, Pack = nb_cigarette 
 	private int weight;
 	private int nb_cigarettes;
-	private int price;
+	private float price;
 	private int tobacco_rate;
 	private int paper_rate;
 	private int agents_rate;
 	private HashMap<String, Float> components;
 	
-	public Pack(Type type, String brand, int weight, int nb_cigarettes, int price, int tobacco_rate, int paper_rate, int agents_rate, HashMap<String, Float> components){
+	public Pack(Type type, String brand, int weight, int nb_cigarettes, float price, int tobacco_rate, int paper_rate, int agents_rate, HashMap<String, Float> components){
 		id = last_id++;
 		this.type = type;
 		this.brand = brand;
